+++ CHANGES.rst
@@ -4,6 +4,11 @@ Changelog
 1.2.3 (unreleased)
 ------------------
 
+- Override Accessibility info with a multilingual / editable / filled by default
+  text. The text can be changed in a new Anysurfer control panel (upgrade step
+  included)
+  [boulch, laulaz]
+
 - Fix 404 / EN pages title handling
   [laulaz]